Refactor beacon_chain.
Moves all the `BeaconChain` impls into the same file that the struct is
defined in.
