+++ beacon_node/beacon_chain/src/attestation_targets.rs
@@ -1,4 +1,3 @@
-use crate::{BeaconChain, ClientDB, SlotClock};
 use std::collections::HashMap;
 use types::Hash256;
 
@@ -19,25 +18,5 @@ impl AttestationTargets {
 
     pub fn insert(&mut self, validator_index: u64, block_hash: Hash256) -> Option<Hash256> {
         self.map.insert(validator_index, block_hash)
-    }
-}
-
-impl<T, U> BeaconChain<T, U>
-where
-    T: ClientDB,
-    U: SlotClock,
-{
-    pub fn insert_latest_attestation_target(&self, validator_index: u64, block_root: Hash256) {
-        let mut targets = self.latest_attestation_targets.write();
-        targets.insert(validator_index, block_root);
-    }
-
-    pub fn get_latest_attestation_target(&self, validator_index: u64) -> Option<Hash256> {
-        let targets = self.latest_attestation_targets.read();
-
-        match targets.get(validator_index) {
-            Some(hash) => Some(hash.clone()),
-            None => None,
-        }
     }
 }
